Module Record_time

A very simple benchmarking API.

val record_time : unit -> unit

By default OCaml uses floats to record dates, which makes me worry about precision. This simple C API does not have this defect. bench_return_time returns the number of microseconds since the last call to bench_record_time.

val return_time : unit -> int